requirements.txt
langchain==0.1.15
langchain-community==0.0.34
langchain-core==0.1.52
langchain-experimental==0.0.57
langchain-google-genai==1.0.1
langchain-groq==0.0.1
langchain-openai==0.0.7
langchain-text-splitters==0.0.1
langchainhub==0.1.15
fastembed==0.2.5
llama-index-core==0.10.25.post1
llama-parse==0.4.0
llamaindex-py-client==0.1.15
qdrant-client==1.8.2
streamlit==1.33.0
streamlit-float==0.3.2
unstructured==0.12.5
unstructured-client==0.18.0
pypdf==4.3.0
python-docx==1.1.2